refactoring names to fileName convention
